+++ src/hera/__init__.py
@@ -1,4 +1,5 @@
 from hera._context import dag_context
+from hera._version import version
 from hera.affinity import (
     Affinity,
     Expression,
@@ -55,3 +56,6 @@ from hera.workflow import Workflow
 from hera.workflow_service import WorkflowService
 from hera.workflow_status import WorkflowStatus
 from hera.workflow_template import WorkflowTemplate
+
+__version__ = version
+__version_info__ = version.split(".")