+++ src/hera/security_context.py
@@ -8,6 +8,7 @@ from argo_workflows.models import Capabilities, PodSecurityContext, SecurityCont
 class BaseSecurityContext:
     """Abstract class to accommodate the shared functionality of task and workflow context."""
 
+    privileged: Optional[bool] = None
     run_as_user: Optional[int] = None
     run_as_group: Optional[int] = None
     run_as_non_root: Optional[bool] = None
@@ -24,6 +25,8 @@ class WorkflowSecurityContext(BaseSecurityContext):
 
     Attributes
     ----------
+    privileged: Optional[bool] = None
+        Allow all the task's container to run as privileged
     run_as_user: Optional[int]
         Sets the user id of the user running in all the containers in the workflow.
     run_as_group: Optional[int]
@@ -48,6 +51,8 @@ class TaskSecurityContext(BaseSecurityContext):
 
     Attributes
     ----------
+    privileged: Optional[bool] = None
+        Allow all the task's container to run as privileged
     run_as_user: Optional[int]
         Sets the user id of the user running in the task's container.
     run_as_group: Optional[int]