+++ src/hera/toleration.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass
 from typing import Optional
 
+from argo_workflows.models import Toleration as ArgoToleration
 
 @dataclass
 class Toleration:
@@ -26,6 +27,9 @@ class Toleration:
     operator: str
     effect: str
     value: Optional[str] = ""
+
+    def to_argo_toleration(self):
+        return ArgoToleration(key=self.key, effect=self.effect, operator=self.operator, value=self.value)
 
 
 GPUToleration = Toleration(key="nvidia.com/gpu", operator="Equal", value="present", effect="NoSchedule")